+++ packages/svelte/src/internal/client/types.d.ts
@@ -99,11 +99,17 @@ export type ComputationSignal<V = unknown> = {
 	/** The types that the signal represent, as a bitwise value */
 	f: SignalFlags;
 	/** init: The function that we invoke for effects and computeds */
-	i: null | (() => V) | (() => void | (() => void)) | ((b: Block) => void | (() => void));
+	i:
+		| null
+		| (() => V)
+		| (() => void | (() => void))
+		| ((b: Block, s: Signal) => void | (() => void));
 	/** references: Anything that a signal owns */
 	r: null | ComputationSignal[];
 	/** value: The latest value for this signal, doubles as the teardown for effects */
 	v: V;
+	/** level: the depth from the root signal, used for ordering render/pre-effects topologically **/
+	l: number;
 };
 
 export type Signal<V = unknown> = SourceSignal<V> | ComputationSignal<V>;
@@ -288,14 +294,7 @@ export type EachBlock = {
 
 export type EachItemBlock = {
 	/** transition */
-	a:
-		| null
-		| ((
-				block: EachItemBlock,
-				transitions: Set<Transition>,
-				index: number,
-				index_is_reactive: boolean
-		  ) => void);
+	a: null | ((block: EachItemBlock, transitions: Set<Transition>) => void);
 	/** dom */
 	d: null | TemplateNode | Array<TemplateNode>;
 	/** effect */