+++ packages/svelte/src/main/ambient.d.ts
@@ -178,4 +178,4 @@ declare function $props<T>(): T;
  */
 declare function $inspect<T extends any[]>(
 	...values: T
-): { with: (type: 'init' | 'update', ...values: T) => void };
+): { with: (fn: (type: 'init' | 'update', ...values: T) => void) => void };