+++ packages/svelte/tests/runtime-runes/samples/event-attribute-delegation-2/main.svelte
@@ -1,0 +1,8 @@
+<script>
+	import { log } from './log.js';
+</script>
+<div on:click={(e) => { log.push('clicked div') }}>
+	<button onclick={(e) => { log.push('clicked button'); e.stopPropagation() }}>
+		Button
+	</button>
+</div>